bricks: Freeze only specific files.
For example, Technic Hub does not need light matrix code.

Also remove single-axis calibration routine. To be replaced with 3D calibration routine.
