+++ bricks/primehub/manifest.py
@@ -1,12 +1,5 @@
-import pathlib
-
 # Common modules
 include("../_common/manifest.py")
 
-# PrimeHub additions, like IMU.
-primehub_modules = list(pathlib.Path("../primehub/modules").glob("*.py"))
-
-if any(primehub_modules):
-    for m in primehub_modules:
-        print(f"Including {m.name} as a module.")
-        freeze_as_mpy(str(m.parent), m.name)
+freeze_as_mpy("../primehub/modules", "_imu_calibrate.py")
+freeze_as_mpy("../primehub/modules", "_light_matrix.py")
