+++ bricks/primehub/modules/_light_matrix.py
@@ -1,0 +1,9 @@
+from pybricks.tools import wait
+
+
+def light_matrix_text_async(display, text, on, off):
+    for char in text:
+        display.char(char)
+        yield from wait(on)
+        display.off()
+        yield from wait(off)
