+++ pybricks/common/pb_type_imu.c
@@ -344,28 +344,6 @@ static mp_obj_t pb_type_imu_reset_heading(size_t n_args, const mp_obj_t *pos_arg
 }
 static MP_DEFINE_CONST_FUN_OBJ_KW(pb_type_imu_reset_heading_obj, 1, pb_type_imu_reset_heading);
 
-// pybricks._common.IMU.update_heading_correction
-static mp_obj_t pb_type_imu_update_heading_correction(mp_obj_t self_in) {
-    pb_type_imu_obj_t *self = MP_OBJ_TO_PTR(self_in);
-    pb_module_tools_assert_blocking();
-
-    // Disable stop button and cache original setting to restore later.
-    pbio_button_flags_t stop_button = pbsys_program_stop_get_buttons();
-
-    nlr_buf_t nlr;
-    if (nlr_push(&nlr) == 0) {
-        mp_obj_t func = pb_function_import_helper(MP_QSTR__hub_extra, MP_QSTR_imu_update_heading_correction);
-        mp_call_function_1(func, self->hub);
-        pbsys_program_stop_set_buttons(stop_button);
-        nlr_pop();
-    } else {
-        pbsys_program_stop_set_buttons(stop_button);
-        nlr_jump(nlr.ret_val);
-    }
-    return mp_const_none;
-}
-MP_DEFINE_CONST_FUN_OBJ_1(pb_type_imu_update_heading_correction_obj, pb_type_imu_update_heading_correction);
-
 // dir(pybricks.common.IMU)
 static const mp_rom_map_elem_t pb_type_imu_locals_dict_table[] = {
     { MP_ROM_QSTR(MP_QSTR_acceleration),     MP_ROM_PTR(&pb_type_imu_acceleration_obj)    },
@@ -379,7 +357,6 @@ static const mp_rom_map_elem_t pb_type_imu_locals_dict_table[] = {
     { MP_ROM_QSTR(MP_QSTR_tilt),             MP_ROM_PTR(&pb_type_imu_tilt_obj)            },
     { MP_ROM_QSTR(MP_QSTR_up),               MP_ROM_PTR(&pb_type_imu_up_obj)              },
     { MP_ROM_QSTR(MP_QSTR_orientation),      MP_ROM_PTR(&common_IMU_orientation_obj)     },
-    { MP_ROM_QSTR(MP_QSTR_update_heading_correction), MP_ROM_PTR(&pb_type_imu_update_heading_correction_obj)},
 };
 static MP_DEFINE_CONST_DICT(pb_type_imu_locals_dict, pb_type_imu_locals_dict_table);
